Move LoginForm into components/forms directory
