+++ src/views/Login.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import LoginForm from '@/components/LoginForm.vue';
+import LoginForm from '@/components/forms/LoginForm.vue';
 
 export default {
   components: {
